pyproject.toml
[project]
name = "doba-s"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorama>=0.4.6",
    "fastapi>=0.115.5",
    "neomodel>=5.4.0",
    "ollama>=0.3.3",
    "openai>=1.52.0",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.32.1",
]
